src/requirements.txt
runpod==1.7.6
transformers==4.41.2
accelerate==0.29.3
safetensors==0.4.3
huggingface_hub==0.23.5
sentencepiece==0.2.0
tokenizers==0.15.2
numpy==1.26.4
pillow==10.3.0
requests==2.32.3

# MoE-LLaVA (официальный репозиторий)
git+https://github.com/PKU-YuanGroup/MoE-LLaVA.git

# Опционально: квантовки
bitsandbytes==0.43.1 ; platform_system == "Linux" and platform_machine == "x86_64"
